[更新]修改更新apk逻辑，现在当下载完成apk后，如果app不在前台，则需要等到app回到前台才会安装新的apk文件

+++ common/library-common/src/main/java/com/donews/common/download/DownloadHelper.java
@@ -8,6 +8,11 @@ import android.util.Log;
 
 import androidx.core.content.FileProvider;
 
+import com.donews.base.base.BaseApplication;
+import com.donews.utilslibrary.utils.AppStatusUtils;
+import com.orhanobut.logger.Logger;
+import com.tencent.mmkv.MMKV;
+
 import java.io.File;
 import java.io.PrintWriter;
 import java.io.StringWriter;
@@ -21,28 +26,65 @@ import java.net.UnknownHostException;
 public class DownloadHelper {
     public static final String PROVIDER_AUTHORITY = ".DNFileProvider";
 
+    public static final String KEY_INSTALL_APK = "installApk";
+    public static final String KEY_APK_PATH = "apkPath";
+
     public static void installApp(Context context, String path) {
-        File file = new File(path);
-        try {
-            if (file.exists()) {
-                Intent intent = new Intent(Intent.ACTION_VIEW);
 
-                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+        if (AppStatusUtils.isForceGround()) {
+            File file = new File(path);
+            try {
+                if (file.exists()) {
+                    Intent intent = new Intent(Intent.ACTION_VIEW);
+                    intent.addCategory("android.intent.category.DEFAULT");
                     intent.setFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
-                    Uri contentUri = FileProvider.getUriForFile(context, context.getPackageName() + PROVIDER_AUTHORITY, file);
                     intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-                    intent.setDataAndType(contentUri, "application/vnd.android.package-archive");
-                } else {
-                    intent.setDataAndType(Uri.fromFile(file), "application/vnd.android.package-archive");
-                    intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+                    Uri uri;
+                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
+                        uri = FileProvider.getUriForFile(context, context.getPackageName() + PROVIDER_AUTHORITY,
+                                file);
+                    } else {
+                        uri = Uri.fromFile(file);
+                    }
+                    intent.setDataAndType(uri, "application/vnd.android.package-archive");
+                    context.startActivity(intent);
+                    clearInstallApp(context, path);
                 }
-                context.startActivity(intent);
+            } catch (Exception e) {
+                Log.e("DownloadHelper", "  DownloadHelper install error: " + getStackTraceString(e));
             }
-        } catch (Exception e) {
-            Log.e("DownloadHelper", "  DownloadHelper install error: " + getStackTraceString(e));
+        } else {
+            saveInstallApp(context, path);
         }
-
     }
+
+    private static void saveInstallApp(Context context, String path) {
+        MMKV mmkv = MMKV.defaultMMKV();
+        if (mmkv != null) {
+            mmkv.encode(KEY_INSTALL_APK, true);
+            mmkv.encode(KEY_APK_PATH, path);
+        }
+    }
+
+    private static void clearInstallApp(Context context, String path) {
+        MMKV mmkv = MMKV.defaultMMKV();
+        if (mmkv != null) {
+            mmkv.encode(KEY_INSTALL_APK, false);
+            mmkv.encode(KEY_APK_PATH, "");
+        }
+    }
+
+    public static void installApp() {
+        MMKV mmkv = MMKV.defaultMMKV();
+        if (mmkv != null) {
+            boolean installApk = mmkv.decodeBool(KEY_INSTALL_APK, false);
+            if (installApk) {
+                String path = mmkv.decodeString(KEY_APK_PATH);
+                installApp(BaseApplication.getInstance().getApplicationContext(), path);
+            }
+        }
+    }
+
 
     protected static String getStackTraceString(Throwable tr) {
         if (tr == null) {
